requirements.txt
fastapi
uvicorn
gradio
python-dotenv
langchain
langchain-openai
openai
pydantic